refactor:LottoRankFinder->WinningLottoTicket으로 이관
