+++ src/test/kotlin/lottoTest/domain/ScoreRankMapTest.kt
@@ -3,14 +3,11 @@ package lottoTest.domain
 import lotto.domain.Lotto
 import lotto.domain.LottoNumber
 import lotto.domain.WinningLottoTicket
-import lotto.service.LottoRankFinder
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 
 class ScoreRankMapTest {
-    private val lottoRankFinder = LottoRankFinder()
-
     @Test
     @DisplayName("복권의 구매액과 총 당첨 정보를 입력받아 수익률을 반환한다")
     fun t1() {
@@ -21,7 +18,7 @@ class ScoreRankMapTest {
             )
         val winningLotto = Lotto(listOf(1, 2, 3, 4, 5, 6).map { LottoNumber.of(it) })
         val bonus = LottoNumber.of(7)
-        val rankMap = lottoRankFinder.findLottoRanks(manyLotto, WinningLottoTicket(winningLotto, bonus))
+        val rankMap = WinningLottoTicket(winningLotto, bonus).findLottoRanks(manyLotto)
         val result = rankMap.getRate()
         assertThat(result).isEqualTo("25.00")
     }
@@ -37,7 +34,7 @@ class ScoreRankMapTest {
             )
         val winningLotto = Lotto(listOf(1, 2, 3, 4, 5, 6).map { LottoNumber.of(it) })
         val bonus = LottoNumber.of(7)
-        val rankMap = lottoRankFinder.findLottoRanks(manyLotto, WinningLottoTicket(winningLotto, bonus))
+        val rankMap = WinningLottoTicket(winningLotto, bonus).findLottoRanks(manyLotto)
         val result = rankMap.getRate()
         assertThat(result).isEqualTo("3.33")
     }
@@ -53,7 +50,7 @@ class ScoreRankMapTest {
             )
         val winningLotto = Lotto(listOf(1, 2, 3, 4, 5, 6).map { LottoNumber.of(it) })
         val bonus = LottoNumber.of(7)
-        val rankMap = lottoRankFinder.findLottoRanks(manyLotto, WinningLottoTicket(winningLotto, bonus))
+        val rankMap = WinningLottoTicket(winningLotto, bonus).findLottoRanks(manyLotto)
         assertThat(rankMap.getEarned()).isEqualTo(10000)
     }
 
@@ -68,7 +65,7 @@ class ScoreRankMapTest {
             )
         val winningLotto = Lotto(listOf(1, 2, 3, 4, 5, 6).map { LottoNumber.of(it) })
         val bonus = LottoNumber.of(7)
-        val rankMap = lottoRankFinder.findLottoRanks(manyLotto, WinningLottoTicket(winningLotto, bonus))
+        val rankMap = WinningLottoTicket(winningLotto, bonus).findLottoRanks(manyLotto)
         assertThat(rankMap.getPaid()).isEqualTo(3000)
     }
 }
